requirements.txt
# Core dependencies
pandas>=2.0.0
openpyxl>=3.1.0
python-dateutil>=2.8.0

# OCR support for scanned PDFs (PRIMARY)
pytesseract>=0.3.10
pdf2image>=1.16.3
Pillow>=10.0.0
PyPDF2>=3.0.0

# Testing (development)
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-mock>=3.11.0

# Code quality (development)
black>=23.7.0
flake8>=6.1.0
mypy>=1.5.0

# Type stubs (development)
pandas-stubs>=2.0.0

# Optional: For native PDF text extraction (not needed for scanned PDFs)
# tabula-py>=2.8.0  # Requires Java and jpype1
# jpype1>=1.4.0     # Complex build dependencies
